+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html). (Patch version X.Y.0 is implied if not specified.)
 
 ## [Unreleased](https://github.com/USGS/time-series-services/compare/time-series-services/compare/0.8.0...master)
+### Added
+-  WaterML can be requested from /collections/{collectionId}/items/{featureId}/observations/statistical-time-series/{timeSeriesId}
 
 ## [0.9.0](https://github.com/USGS/time-series-services/compare/time-series-services-0.8.0...time-series-services-0.9.0) - 2021-01-11
 ### Changed